Add browser-based interface for reading dump QR codes from a video feed
In practice the video feed may be from a capture card or a camera (handled the same by the os/browser)

Using javascript, Emscripten, wasm, zbar

file:// cannot be used because the whole thing is split into several files.
http:// cannot be used because that prevents accessing video feeds.
So https:// must be used. The https_server python module provides one, and has a comment for generating a required (self-signed) certificate.

Building and running:
1) build zbar with Emscripten: run build_zbar.sh
2) build the "website": run build.sh
3) visit webdecoder/website/main.html (not webdecoder/main.html !)

The generated "website" folder (by build.sh) can be copied elsewhere (e.g. on a server)